requirements.txt
openai
gradio
python-dotenv
pydub
pillow
